# pkg/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: Sandbox0 API
  version: 0.1.0
  description: Public HTTP APIs exposed by internal-gateway.
servers:
  - url: https://api.sandbox0.ai
    description: Production
...
paths:
...
  /api/v1/sandboxes:
    post:
      tags: [sandboxes]
      summary: Create (claim) a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequest"
      responses:
        "201":
          description: Sandbox claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessClaimResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/sandboxes/{id}:
    get:
      tags: [sandboxes]
      summary: Get sandbox by ID
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [sandboxes]
      summary: Update sandbox configuration
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxUpdateRequest"
      responses:
        "200":
          description: Sandbox updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSandboxResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [sandboxes]
      summary: Delete (terminate) a sandbox
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/sandboxes/{id}
      parameters:
        - $ref: "#/components/parameters/SandboxID"
      responses:
        "200":
          description: Sandbox terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/templates:
    get:
      tags: [templates]
      summary: List templates
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateListResponse"
    post:
      tags: [templates]
      summary: Create template
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateCreateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateResponse"
  /api/v1/templates/{id}:
    get:
      tags: [templates]
      summary: Get template
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates/{id}
      parameters:
        - $ref: "#/components/parameters/TemplateID"
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [templates]
      summary: Update template
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates/{id}
      parameters:
        - $ref: "#/components/parameters/TemplateID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdateRequest"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTemplateResponse"
    delete:
      tags: [templates]
      summary: Delete template
      security:
        - bearerAuth: []
      x-upstream-service: manager
      x-upstream-path: /api/v1/templates/{id}
      parameters:
        - $ref: "#/components/parameters/TemplateID"
      responses:
        "200":
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SandboxID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    TemplateID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
      required: [code, message]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/Error"
      required: [success, error]
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data: {}
      required: [success]
...
    SuccessMessageResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
...
    ClaimRequest:
      type: object
      properties:
        template:
          type: string
        config:
          $ref: "#/components/schemas/SandboxConfig"
      required: []
    SandboxConfig:
      type: object
      properties:
        env_vars:
          type: object
          additionalProperties:
            type: string
        ttl:
          type: integer
          format: int32
        hard_ttl:
          type: integer
          format: int32
        network:
          $ref: "#/components/schemas/TplSandboxNetworkPolicy"
        webhook:
          $ref: "#/components/schemas/WebhookConfig"
        auto_resume:
          type: boolean
          default: false
        exposed_ports:
          type: array
          items:
            $ref: "#/components/schemas/ExposedPortConfig"
    ExposedPortConfig:
      type: object
      properties:
        port:
          type: integer
          format: int32
        auto_wakeup:
          type: boolean
      required: [port, auto_wakeup]
    WebhookConfig:
      type: object
      properties:
        url:
          description: Required when webhook is enabled. Target URL that receives event callbacks.
          type: string
        secret:
          description: Optional. Shared secret used to sign webhook payloads.
          type: string
        watch_dir:
          description: Optional. When set, procd subscribes to file events under this directory (same semantics as the file watch WebSocket API) and emits file.modified events.
          type: string
    ClaimResponse:
      type: object
      properties:
        sandbox_id:
          type: string
        status:
          type: string
        pod_name:
          type: string
        template:
          type: string
        cluster_id:
          type: string
          nullable: true
      required: [sandbox_id, status, pod_name, template]
    SuccessClaimResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ClaimResponse"
    Sandbox:
      type: object
      properties:
        id:
          type: string
        template_id:
          type: string
        team_id:
          type: string
        user_id:
          type: string
        status:
          type: string
        paused:
          type: boolean
        auto_resume:
          type: boolean
        exposed_ports:
          type: array
          items:
            $ref: "#/components/schemas/ExposedPortConfig"
        pod_name:
          type: string
        expires_at:
          type: string
          format: date-time
        claimed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - id
        - template_id
        - team_id
        - status
        - paused
        - auto_resume
        - pod_name
        - expires_at
        - claimed_at
        - created_at
    SandboxUpdateRequest:
      type: object
      properties:
        config:
          $ref: "#/components/schemas/SandboxConfig"
...
    SuccessSandboxResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Sandbox"
...
    TemplateCreateRequest:
      type: object
      properties:
        template_id:
          type: string
        spec:
          $ref: "#/components/schemas/SandboxTemplateSpec"
      required: [template_id, spec]
    TemplateUpdateRequest:
      type: object
      properties:
        spec:
          $ref: "#/components/schemas/SandboxTemplateSpec"
      required: [spec]
    Template:
      type: object
      properties:
        template_id:
          type: string
        scope:
          type: string
        team_id:
          type: string
        user_id:
          type: string
        spec:
          $ref: "#/components/schemas/SandboxTemplateSpec"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [template_id, scope, spec, created_at, updated_at]
    SuccessTemplateResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Template"
    SuccessTemplateListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                templates:
                  type: array
                  items:
                    $ref: "#/components/schemas/Template"
                count:
                  type: integer
...
    SandboxTemplateSpec:
      type: object
      properties:
        description:
          type: string
        displayName:
          type: string
        tags:
          type: array
          items:
            type: string
        mainContainer:
          $ref: "#/components/schemas/ContainerSpec"
        sidecars:
          type: array
          items:
            $ref: "#/components/schemas/ContainerSpec"
        pod:
          $ref: "#/components/schemas/PodSpecOverride"
        network:
          $ref: "#/components/schemas/TplSandboxNetworkPolicy"
        pool:
          $ref: "#/components/schemas/PoolStrategy"
        lifecycle:
          $ref: "#/components/schemas/LifecyclePolicy"
        envVars:
          type: object
          additionalProperties:
            type: string
        public:
          type: boolean
        allowedTeams:
          type: array
          items:
            type: string
        runtimeClassName:
          type: string
        clusterId:
          type: string
    ContainerSpec:
      type: object
      properties:
        image:
          type: string
        imagePullPolicy:
          type: string
        env:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
        resources:
          $ref: "#/components/schemas/ResourceQuota"
        securityContext:
          $ref: "#/components/schemas/SecurityContext"
      required: [image, resources]
    EnvVar:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
      required: [name, value]
    ResourceQuota:
      type: object
      properties:
        cpu:
          type: string
        memory:
          type: string
    SecurityContext:
      type: object
      properties:
        capabilities:
          $ref: "#/components/schemas/Capabilities"
        runAsUser:
          type: integer
          format: int64
        runAsGroup:
          type: integer
          format: int64
    Capabilities:
      type: object
      properties:
        drop:
          type: array
          items:
            type: string
    PodSpecOverride:
      type: object
      properties:
        nodeSelector:
          type: object
          additionalProperties:
            type: string
        affinity:
          $ref: "#/components/schemas/Affinity"
        tolerations:
          type: array
          items:
            $ref: "#/components/schemas/Toleration"
        serviceAccountName:
          type: string
    Affinity:
      type: object
      properties:
        nodeAffinity:
          $ref: "#/components/schemas/NodeAffinity"
        podAffinity:
          $ref: "#/components/schemas/PodAffinity"
    NodeAffinity:
      type: object
      properties:
        requiredDuringSchedulingIgnoredDuringExecution:
          $ref: "#/components/schemas/NodeSelector"
        preferredDuringSchedulingIgnoredDuringExecution:
          type: array
          items:
            $ref: "#/components/schemas/PreferredSchedulingTerm"
    NodeSelector:
      type: object
      properties:
        nodeSelectorTerms:
          type: array
          items:
            $ref: "#/components/schemas/NodeSelectorTerm"
    NodeSelectorTerm:
      type: object
      properties:
        matchExpressions:
          type: array
          items:
            $ref: "#/components/schemas/NodeSelectorRequirement"
        matchFields:
          type: array
          items:
            $ref: "#/components/schemas/NodeSelectorRequirement"
    NodeSelectorRequirement:
      type: object
      properties:
        key:
          type: string
        operator:
          type: string
        values:
          type: array
          items:
            type: string
      required: [key, operator]
    PreferredSchedulingTerm:
      type: object
      properties:
        weight:
          type: integer
          format: int32
        preference:
          $ref: "#/components/schemas/NodeSelectorTerm"
      required: [weight, preference]
    PodAffinity:
      type: object
      properties:
        requiredDuringSchedulingIgnoredDuringExecution:
          type: array
          items:
            $ref: "#/components/schemas/PodAffinityTerm"
        preferredDuringSchedulingIgnoredDuringExecution:
          type: array
          items:
            $ref: "#/components/schemas/WeightedPodAffinityTerm"
    PodAffinityTerm:
      type: object
      properties:
        labelSelector:
          $ref: "#/components/schemas/LabelSelector"
        namespaces:
          type: array
          items:
            type: string
        topologyKey:
          type: string
      required: [topologyKey]
    LabelSelector:
      type: object
      properties:
        matchLabels:
          type: object
          additionalProperties:
            type: string
        matchExpressions:
          type: array
          items:
            $ref: "#/components/schemas/LabelSelectorRequirement"
    LabelSelectorRequirement:
      type: object
      properties:
        key:
          type: string
        operator:
          type: string
        values:
          type: array
          items:
            type: string
      required: [key, operator]
    WeightedPodAffinityTerm:
      type: object
      properties:
        weight:
          type: integer
          format: int32
        podAffinityTerm:
          $ref: "#/components/schemas/PodAffinityTerm"
      required: [weight, podAffinityTerm]
    Toleration:
      type: object
      properties:
        key:
          type: string
        operator:
          type: string
        value:
          type: string
        effect:
          type: string
    PoolStrategy:
      type: object
      properties:
        minIdle:
          type: integer
          format: int32
        maxIdle:
          type: integer
          format: int32
        autoScale:
          type: boolean
      required: [minIdle, maxIdle, autoScale]
    TplSandboxNetworkPolicy:
      type: object
      properties:
        mode:
          type: string
          enum: [allow-all, block-all]
        egress:
          $ref: "#/components/schemas/NetworkEgressPolicy"
      required: [mode]
    NetworkEgressPolicy:
      type: object
      properties:
        allowedCidrs:
          type: array
          items:
            type: string
        allowedDomains:
          type: array
          items:
            type: string
        allowedPorts:
          type: array
          items:
            $ref: "#/components/schemas/PortSpec"
        deniedDomains:
          type: array
          items:
            type: string
        deniedCidrs:
          type: array
          items:
            type: string
        deniedPorts:
          type: array
          items:
            $ref: "#/components/schemas/PortSpec"
    PortSpec:
      type: object
      properties:
        port:
          type: integer
          format: int32
        protocol:
          type: string
        endPort:
          type: integer
          format: int32
      required: [port]
    LifecyclePolicy:
      type: object
      properties:
        defaultTTL:
          type: integer
          format: int32
        maxTTL:
          type: integer
          format: int32
        idleTimeout:
          type: integer
          format: int32
        preStop:
          $ref: "#/components/schemas/PreStopHook"
    PreStopHook:
      type: object
      properties:
        command:
          type: array
          items:
            type: string
        timeoutSeconds:
          type: integer
          format: int32
...